requirements.txt
web3
python-dotenv
jsonlib
pandas
asyncio
db-sqlite3
pytest
DateTime
python-time
pycoingecko
aiohttp
requests
python-csv
fastapi 
pydantic
